Add operation to load features from GeoJSON.
This commit adds an operation to load features from GeoJSON and
accompanying tests.

The GeoJSON can be either just the geometry clause of the GeoJSON
(as sometimes seen in the wild) or the full feature specification.

The loaded features have the data type Option[JsonNode], where JsonNode
is from Jackson and can be transformed or accessed by custom code into
whatever data representation is necessary.

The GeoJSON parser does not currently support GeometryCollections or
FeatureCollections.